+++ ouds_theme_orange/lib/orange_theme.dart
@@ -19,11 +19,13 @@ import 'package:ouds_theme_contract/theme/tokens/components/ouds_components_toke
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_border_semantic_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_color_semantic_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_font_semantic_tokens.dart';
+import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_opacity_semantic_tokens.dart';
 import 'package:ouds_theme_orange/material/orange_material_color_dark_tokens.dart';
 import 'package:ouds_theme_orange/material/orange_material_color_light_tokens.dart';
 import 'package:ouds_theme_orange/semantic/orange_border_semantic_tokens.dart';
 import 'package:ouds_theme_orange/semantic/orange_color_semantic_tokens.dart';
 import 'package:ouds_theme_orange/semantic/orange_font_semantic_tokens.dart';
+import 'package:ouds_theme_orange/semantic/orange_opacity_semantic_tokens.dart';
 
 /// [OrangeTheme] implements the [OudsThemeContract] and defines the "Orange" theme for the application.
 /// It provides both light and dark theme configurations using the OUDS design system, including color, typography,
@@ -66,6 +68,9 @@ class OrangeTheme implements OudsThemeContract {
       OrangeColorSemanticTokens();
 
   @override
+  OudsOpacitySemanticTokens get opacityTokens => OrangeOpacitySemanticTokens();
+
+  @override
   OudsBorderSemanticTokens get borderTokens => OrangeBorderSemanticTokens();
 
   @override
@@ -78,6 +83,7 @@ class OrangeTheme implements OudsThemeContract {
   @override
   OudsProvidersTokens get providersTokens => OudsProvidersTokens(
       colorScheme: colorsScheme,
+      opacityTokens: opacityTokens,
       borderTokens: borderTokens,
       fontTokens: fontTokens);