+++ ouds_theme_white_label/lib/white_label_theme.dart
@@ -20,11 +20,13 @@ import 'package:ouds_theme_contract/theme/tokens/components/ouds_components_toke
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_border_semantic_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_color_semantic_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_font_semantic_tokens.dart';
+import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_opacity_semantic_tokens.dart';
 import 'package:ouds_theme_white_label/material/white_label_material_color_dark_tokens.dart';
 import 'package:ouds_theme_white_label/material/white_label_material_color_light_tokens.dart';
 import 'package:ouds_theme_white_label/semantic/white_label_border_semantic_tokens.dart';
 import 'package:ouds_theme_white_label/semantic/white_label_color_semantic_tokens.dart';
 import 'package:ouds_theme_white_label/semantic/white_label_font_semantic_tokens.dart';
+import 'package:ouds_theme_white_label/semantic/white_label_opacity_semantic_tokens.dart';
 
 /// [WhiteLabelTheme] implements the [OudsThemeContract] and defines the theme for the "White label" version of the application.
 /// This theme provides configurations for both light and dark modes, with customizable color schemes, typography,
@@ -64,6 +66,10 @@ class WhiteLabelTheme implements OudsThemeContract {
       WhiteLabelColorSemanticTokens();
 
   @override
+  OudsOpacitySemanticTokens get opacityTokens =>
+      WhiteLabelOpacitySemanticTokens();
+
+  @override
   OudsBorderSemanticTokens get borderTokens => WhiteLabelBorderSemanticTokens();
 
   @override
@@ -78,6 +84,7 @@ class WhiteLabelTheme implements OudsThemeContract {
   @override
   OudsProvidersTokens get providersTokens => OudsProvidersTokens(
       colorScheme: colorsScheme,
+      opacityTokens: opacityTokens,
       borderTokens: borderTokens,
       fontTokens: fontTokens);